Consolidated LDP imports into `ldp_shims` module (#772)
